backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
pandas
